requirements.txt
fastapi==0.115.12
pydantic==2.11.5
requests==2.32.4
pydantic-settings==2.9.1
boto3==1.38.36
cachetools==6.0.0
pyjwt==2.10.1
passlib==1.7.4
uvicorn==0.34.3
bcrypt==4.3.0
python-multipart==0.0.20
httpx==0.28.1